Remove precertificates.go
