+++ sa/saro.go
@@ -375,6 +375,43 @@ func ReverseName(domain string) string {
 	return strings.Join(labels, ".")
 }
 
+// GetSerialMetadata returns metadata stored alongside the serial number,
+// such as the RegID whose certificate request created that serial, and when
+// the certificate with that serial will expire.
+func (ssa *SQLStorageAuthorityRO) GetSerialMetadata(ctx context.Context, req *sapb.Serial) (*sapb.SerialMetadata, error) {
+	if req == nil || req.Serial == "" {
+		return nil, errIncompleteRequest
+	}
+
+	if !core.ValidSerial(req.Serial) {
+		return nil, fmt.Errorf("invalid serial %q", req.Serial)
+	}
+
+	recordedSerial := recordedSerialModel{}
+	err := ssa.dbReadOnlyMap.WithContext(ctx).SelectOne(
+		&recordedSerial,
+		"SELECT * FROM serials WHERE serial = ?",
+		req.Serial,
+	)
+	if err != nil {
+		if db.IsNoRows(err) {
+			return nil, berrors.NotFoundError("serial %q not found", req.Serial)
+		}
+		return nil, err
+	}
+
+	return &sapb.SerialMetadata{
+		Serial:         recordedSerial.Serial,
+		RegistrationID: recordedSerial.RegistrationID,
+		Created:        recordedSerial.Created.UnixNano(),
+		Expires:        recordedSerial.Expires.UnixNano(),
+	}, nil
+}
+
+func (ssa *SQLStorageAuthority) GetSerialMetadata(ctx context.Context, req *sapb.Serial) (*sapb.SerialMetadata, error) {
+	return ssa.SQLStorageAuthorityRO.GetSerialMetadata(ctx, req)
+}
+
 // GetCertificate takes a serial number and returns the corresponding
 // certificate, or error if it does not exist.
 func (ssa *SQLStorageAuthorityRO) GetCertificate(ctx context.Context, req *sapb.Serial) (*corepb.Certificate, error) {
